requirements.txt
aiofiles==0.4.0
certifi==2019.9.11
chardet==3.0.4
httpcore==0.3.0
httptools==0.0.13
idna==2.8
np==1.0.2
numpy==1.17.3
python-chess==0.28.3
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.9.0
ujson==1.35
urllib3==1.25.6
uvloop==0.13.0
websockets==8.0.2
